+++ pasta_eln/GUI/data_hierarchy/delete_column_delegate.py
@@ -7,9 +7,9 @@
 #  Filename: delete_column_delegate.py
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
-
-import logging
 from typing import Union
+import qtawesome as qta
+from pandas import DataFrame
 from PySide6.QtCore import QAbstractItemModel, QEvent, QModelIndex, QPersistentModelIndex, QSize, Signal
 from PySide6.QtGui import QPainter
 from PySide6.QtWidgets import (QApplication, QPushButton, QStyle, QStyledItemDelegate, QStyleOption, QStyleOptionButton,
@@ -24,12 +24,18 @@ class DeleteColumnDelegate(QStyledItemDelegate):
   delete_clicked_signal = Signal(
     int)  # Signal to inform the delete button click with the position in the table as the parameter
 
-  def __init__(self) -> None:
+  def __init__(self, df:DataFrame, group:str) -> None:
     """
       Constructor
+
+    Args:
+      df (DataFrame): pandas dataframe containing the entire schema
+      group (str): string of this group/class
     """
     super().__init__()
-    self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+    self.df = df
+    self.group = group
+
 
   def paint(self,
             painter: QPainter,
@@ -41,18 +47,21 @@ class DeleteColumnDelegate(QStyledItemDelegate):
       painter (QPainter): Painter instance for painting the button.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
       index (Union[QModelIndex, QPersistentModelIndex]): Cell index.
-
-    Returns: None
-
     """
+    dfSub = self.df[self.df['class']==self.group]
+    if index.row()>=len(dfSub):
+      return
+    if self.group=='' and dfSub[dfSub['idx']==index.row()]['name'].isin(['name','tags','comment']).values[0]:
+      return
     button = QPushButton()
     opt = QStyleOptionButton()
     opt.state = QStyle.StateFlag.State_Active | QStyle.StateFlag.State_Enabled  # type: ignore[attr-defined]
-    opt.rect = option.rect  # type: ignore[attr-defined]
-    opt.icon = QApplication.style().standardIcon(  # type: ignore[attr-defined]
-      QStyle.StandardPixmap.SP_DialogDiscardButton)
-    opt.iconSize = QSize(15, 15)  # type: ignore[attr-defined]
+    opt.rect = option.rect                                                      # type: ignore[attr-defined]
+    opt.icon = qta.icon('fa5s.trash', scale_factor=1.0)                         # type: ignore[attr-defined]
+    opt.iconSize = QSize(15, 15)                                                # type: ignore[attr-defined]
     QApplication.style().drawControl(QStyle.ControlElement.CE_PushButton, opt, painter, button)
+    return
+
 
   def createEditor(self,
                    parent: QWidget,
@@ -64,11 +73,9 @@ class DeleteColumnDelegate(QStyledItemDelegate):
       parent (QWidget): Parent table view.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
       index (Union[QModelIndex, QPersistentModelIndex]): Cell index.
-
-    Returns: None
-
     """
     return None  # type: ignore[return-value]
+
 
   def editorEvent(self,
                   event: QEvent,
@@ -84,11 +91,13 @@ class DeleteColumnDelegate(QStyledItemDelegate):
       index (Union[QModelIndex, QPersistentModelIndex]): Table cell index.
 
     Returns (bool): True if deleted otherwise False
-
     """
+    dfSub = self.df[self.df['class']==self.group]
+    if index.row()>=len(dfSub):
+      return False
+    if self.group=='' and dfSub[dfSub['idx']==index.row()]['name'].isin(['name','tags','comment']).values[0]:
+      return False
     if is_click_within_bounds(event, option):
-      row = index.row()
-      self.logger.info('Delete signal emitted for the position: {%s}', row)
-      self.delete_clicked_signal.emit(row)
+      self.delete_clicked_signal.emit(index.row())
       return True
     return False