+++ pasta_eln/GUI/data_hierarchy/mandatory_column_delegate.py
@@ -8,8 +8,8 @@
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
 
-import logging
 from typing import Union
+from pandas import DataFrame
 from PySide6.QtCore import QAbstractItemModel, QEvent, QModelIndex, QPersistentModelIndex, QRect, Qt
 from PySide6.QtGui import QPainter
 from PySide6.QtWidgets import (QApplication, QRadioButton, QStyle, QStyledItemDelegate, QStyleOptionButton,
@@ -21,12 +21,17 @@ class MandatoryColumnDelegate(QStyledItemDelegate):
   Delegate for creating the radio buttons for the mandatory column in data hierarchy editor tables
   """
 
-  def __init__(self) -> None:
+  def __init__(self, df:DataFrame, group:str) -> None:
     """
       Constructor
+
+    Args:
+      df (DataFrame): pandas dataframe containing the entire schema
+      group (str): string of this group/class
     """
     super().__init__()
-    self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+    self.df = df
+    self.group = group
 
   def paint(self,
             painter: QPainter,
@@ -51,9 +56,14 @@ class MandatoryColumnDelegate(QStyledItemDelegate):
                      option_rect.top(),
                      option_rect.width(),
                      option_rect.height())
-    is_mandatory = bool(index.data(Qt.ItemDataRole.UserRole))
-    opt.state = QStyle.StateFlag.State_On if is_mandatory else QStyle.StateFlag.State_Off  # type: ignore[attr-defined]
+    dfSub = self.df[self.df['class']==self.group]
+    if index.row()>len(dfSub):
+      return
+    trues = dfSub[dfSub['idx']==index.row()]['mandatory'].values
+    isMandatory = len(trues)==1 and trues[0]=='T'
+    opt.state = QStyle.StateFlag.State_On if isMandatory else QStyle.StateFlag.State_Off  # type: ignore[attr-defined]
     style.drawControl(QStyle.ControlElement.CE_RadioButton, opt, painter, radio_button)
+
 
   def editorEvent(self,
                   event: QEvent,
@@ -69,11 +79,14 @@ class MandatoryColumnDelegate(QStyledItemDelegate):
       index (Union[QModelIndex, QPersistentModelIndex]): Table cell index.
 
     Returns (bool): True/False
-
     """
     if event.type() == QEvent.Type.MouseButtonRelease:
-      model.setData(index, not bool(index.data(Qt.ItemDataRole.UserRole)), Qt.ItemDataRole.UserRole)
+      column= (self.df['class']==self.group) & (self.df['idx']==index.row())
+      trues = self.df[column]['mandatory'].values
+      isMandatory = len(trues)==1 and trues[0]=='T'
+      self.df.loc[column,   'mandatory']= '' if isMandatory else 'T'
     return super().editorEvent(event, model, option, index)
+
 
   def createEditor(self,
                    parent: QWidget,
@@ -87,6 +100,5 @@ class MandatoryColumnDelegate(QStyledItemDelegate):
       index (Union[QModelIndex, QPersistentModelIndex]): Cell index.
 
     Returns: None
-
     """
     return None  # type: ignore[return-value]