+++ pasta_eln/GUI/data_hierarchy/reorder_column_delegate.py
@@ -7,9 +7,9 @@
 #  Filename: reorder_column_delegate.py
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
-
-import logging
 from typing import Union
+import qtawesome as qta
+from pandas import DataFrame
 from PySide6.QtCore import QAbstractItemModel, QEvent, QModelIndex, QPersistentModelIndex, QSize, Signal
 from PySide6.QtGui import QPainter
 from PySide6.QtWidgets import (QApplication, QPushButton, QStyle, QStyledItemDelegate, QStyleOptionButton,
@@ -23,12 +23,18 @@ class ReorderColumnDelegate(QStyledItemDelegate):
   """
   re_order_signal = Signal(int)
 
-  def __init__(self) -> None:
+  def __init__(self, df:DataFrame, group:str) -> None:
     """
       Constructor
+
+    Args:
+      df (DataFrame): pandas dataframe containing the entire schema
+      group (str): string of this group/class
     """
     super().__init__()
-    self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+    self.df = df
+    self.group = group
+
 
   def paint(self,
             painter: QPainter,
@@ -40,17 +46,18 @@ class ReorderColumnDelegate(QStyledItemDelegate):
       painter (QPainter): Painter instance for painting the button.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
       index (Union[QModelIndex, QPersistentModelIndex]): Table cell index
-
-    Returns: None
-
     """
+    if index.row()>=len(self.df[self.df['class']==self.group]):
+      return
     button = QPushButton()
     opt = QStyleOptionButton()
     opt.state = QStyle.StateFlag.State_Active | QStyle.StateFlag.State_Enabled  # type: ignore[attr-defined]
-    opt.rect = option.rect  # type: ignore[attr-defined]
-    opt.icon = QApplication.style().standardIcon(QStyle.StandardPixmap.SP_ArrowUp)  # type: ignore[attr-defined]
-    opt.iconSize = QSize(15, 15)  # type: ignore[attr-defined]
+    opt.rect = option.rect                                                      # type: ignore[attr-defined]
+    opt.icon = qta.icon('fa5s.arrow-up', scale_factor=1.0)                      # type: ignore[attr-defined]
+    opt.iconSize = QSize(15, 15)                                                # type: ignore[attr-defined]
     QApplication.style().drawControl(QStyle.ControlElement.CE_PushButton, opt, painter, button)
+    return
+
 
   def createEditor(self,
                    parent: QWidget,
@@ -62,11 +69,9 @@ class ReorderColumnDelegate(QStyledItemDelegate):
       parent (QWidget): Parent table view.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
       index (Union[QModelIndex, QPersistentModelIndex]): Cell index.
-
-    Returns: None
-
     """
     return None  # type: ignore[return-value]
+
 
   def editorEvent(self,
                   event: QEvent,
@@ -82,11 +87,8 @@ class ReorderColumnDelegate(QStyledItemDelegate):
       index (Union[QModelIndex, QPersistentModelIndex]): Table cell index.
 
     Returns (bool): True/False
-
     """
     if is_click_within_bounds(event, option):
-      row = index.row()
-      self.logger.info('Re-order signal emitted for the position: {%s} in the table..', row)
-      self.re_order_signal.emit(row)
+      self.re_order_signal.emit(index.row())
       return True
     return False