+++ pasta_eln/webclient/utils.py
@@ -8,15 +8,14 @@
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
 
-from asyncio import CancelledError, IncompleteReadError, InvalidStateError, LimitOverrunError
-from asyncio import TimeoutError as AsyncTimeoutError
+from asyncio import (CancelledError, IncompleteReadError, InvalidStateError, LimitOverrunError,
+                     TimeoutError as AsyncTimeoutError)
 from functools import wraps
 from json import JSONDecodeError
 from typing import Any, Callable
 from xml.etree.ElementTree import ParseError
 from aiohttp import ClientConnectorError, InvalidURL
-from requests.exceptions import ConnectionError as RequestsConnectionError
-from requests.exceptions import InvalidSchema, MissingSchema
+from requests.exceptions import ConnectionError as RequestsConnectionError, InvalidSchema, MissingSchema
 
 
 def handle_dataverse_exception_async(wrapped: Callable[..., Any]) -> Callable[..., Any]: